+++ example/src/App.tsx
@@ -18,6 +18,11 @@ function App(): React.JSX.Element {
   const [subscribeToStreams, setSubscribeToStreams] =
     React.useState<boolean>(true);
   const [streamProperties, setStreamProperties] = React.useState<Any>({});
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const [maxVideoBitrate, setMaxVideoBitrate] = React.useState<number>(0);
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const [videoBitratePreset, setVideoBitratePreset] =
+    React.useState<string>('bw_saver');
 
   const sessionRef = useRef<OTSession>(null);
   const subscriberRef = useRef<OTSubscriber>(null);
@@ -56,7 +61,8 @@ function App(): React.JSX.Element {
               type: 'greeting2',
               data: 'hello again from React Native',
             });
-            sessionRef.current?.getCapabilities()
+            sessionRef.current
+              ?.getCapabilities()
               .then((capabilities) =>
                 console.log('capabilities:', capabilities)
               );
@@ -87,7 +93,8 @@ function App(): React.JSX.Element {
               // sessionRef.current?.forceDisconnect(event.connectionId);
             }, 5000);
           },
-          connectionDestroyed: (event: any) => console.log('connectionDestroyed', event),
+          connectionDestroyed: (event: any) =>
+            console.log('connectionDestroyed', event),
           archiveStarted: (event: any) =>
             console.log('archiveStarted event', event),
           archiveStopped: (event: any) =>
@@ -119,14 +126,18 @@ function App(): React.JSX.Element {
               // enableDtx: true,
               name: 'OTRN',
               // videoContentHint: 'text',
+              // maxVideoBitrate,
+              // videoBitratePreset,
             }}
             eventHandlers={{
               error: (event: any) => console.log('pub error', event),
               streamCreated: (event: any) => {
                 console.log('pub streamCreated', event);
                 setTimeout(() => {
-                  publisherRef.current?.getRtcStatsReport();
-                }, 4000);
+                  // publisherRef.current?.getRtcStatsReport();
+                  setMaxVideoBitrate(2000000);
+                  setVideoBitratePreset('extra_bw_saver');
+                }, 10000);
               },
               streamDestroyed: (event: any) =>
                 console.log('pub streamDestroyed', event),