+++ ios/OTRNPublisherComponentView.mm
@@ -49,6 +49,8 @@ using namespace facebook::react;
         @"publishVideo" : @(viewProps.publishVideo),
         @"publishCaptions" : @(viewProps.publishCaptions),
         @"allowAudioCaptureWhileMuted" : @(viewProps.allowAudioCaptureWhileMuted),
+        @"maxVideoBitrate" : @(viewProps.maxVideoBitrate),
+        @"videoBitratePreset" : RCTNSStringFromString(viewProps.videoBitratePreset),
     };
 }
 
@@ -96,6 +98,14 @@ using namespace facebook::react;
 
     if (oldViewProps.videoContentHint != newViewProps.videoContentHint) {
         [_impl setVideoContentHint:RCTNSStringFromString(newViewProps.videoContentHint)];
+    }
+
+    if (oldViewProps.maxVideoBitrate != newViewProps.maxVideoBitrate) {
+        [_impl setMaxVideoBitrate:(newViewProps.maxVideoBitrate)];
+    }
+
+    if (oldViewProps.videoBitratePreset != newViewProps.videoBitratePreset) {
+        [_impl setVideoBitratePreset:RCTNSStringFromString(newViewProps.videoBitratePreset)];
     }
 
     if (oldViewProps.cameraTorch != newViewProps.cameraTorch) {