+++ ios/OTRNPublisherComponentView.swift
@@ -150,6 +150,18 @@ import React
             properties["publishCaptions"] as Any
         )
 
+        if let maxVideoBitrate = properties["maxVideoBitrate"] as? Int32 {
+            if (maxVideoBitrate != 0) {
+                publisher.maxVideoBitrate = maxVideoBitrate
+            }
+        }
+
+ 	       if let videoBitratePreset = properties["videoBitratePreset"] as? String {
+            if (videoBitratePreset != "") {
+                publisher.videoBitratePreset = Utils.convertVideoBitratePreset(videoBitratePreset)
+            }
+        }
+
         if let pubView = publisher.view {
             pubView.frame = strictUIViewContainer?.bounds ?? .zero
             publisherUIView = pubView
@@ -263,6 +275,51 @@ import React
 
         publisher.videoCapture?.videoContentHint =
             Utils.convertVideoContentHint(videoContentHint)
+    }
+
+    @objc public func setMaxVideoBitrate(_ maxVideoBitrate: Int32) {
+        guard let publisherId = self.publisherId else {
+            strictUIViewContainer?.handleError([
+                "code": "OTPublisherError",
+                "message": "Publisher ID is not set",
+            ])
+            return
+        }
+
+        guard let publisher = OTRN.sharedState.publishers[publisherId] else {
+            strictUIViewContainer?.handleError([
+                "code": "OTPublisherError",
+                "message": "Could not find publisher instance",
+            ])
+            return
+        }
+
+        if (maxVideoBitrate != 0) {
+            publisher.maxVideoBitrate = maxVideoBitrate
+        }
+    }
+
+    @objc public func setVideoBitratePreset(_ videoBitratePreset: String) {
+        guard let publisherId = self.publisherId else {
+            strictUIViewContainer?.handleError([
+                "code": "OTPublisherError",
+                "message": "Publisher ID is not set",
+            ])
+            return
+        }
+
+        guard let publisher = OTRN.sharedState.publishers[publisherId] else {
+            strictUIViewContainer?.handleError([
+                "code": "OTPublisherError",
+                "message": "Could not find publisher instance",
+            ])
+            return
+        }
+
+        if (videoBitratePreset != "") {
+            publisher.videoBitratePreset = 
+                Utils.convertVideoBitratePreset(videoBitratePreset)
+        }
     }
 
     @objc public func cleanup() {